fix(shared-source/annotate_rsids_from_linker.cc): add manual exception for colliding pseudogene rsID
downstream tests find a pair of rsIDs (rs1160972848, rs1192281978) that point to two variants at
different locations *on the dbSNP website* but not *in the dbSNP database* where only one of the two
rsIDs points to both of them. the TOPMed allele frequencies are identical but the association
results are different. this is significantly beyond the scope of this pipeline. since I've only
found one exception here, I've added a single manual override for these IDs that replaces them with
their chr:pos:ref:alt IDs. if other problematic variants become apparent, or if I find myself with
an abundance of time, I'll add in a more rigorous and transparent fix at a later date

re #28