feat: add unintegrated module for handing psa residual output files
based on how many random side projects are getting requested through the "atlas" system, there is an
increasing need for a way to specify output formatting in a user-config-level and modular way. this
was the first swing at that, formatting the preprocessed PSA residual analyses for Mitch Machiela
and Sonja Berndt. initially the goal was to create an entire postprocessing interface in config and
handle it there, but that process was taking forever to even conceptualize, so instead this is just
a standalone that I ran separately. at some future date, a configuration option for specifying which
postprocessing module should be applied to a file needs to be added. also at some point, I don't
know if this is just a Sonja thing, but these requests keep specifically wanting the minimac4 native
Rsq values, which aren't pulled in by the pipeline normally; that might need to be added as a
central feature someday.

re #33