feat: add testing addition of external combine_categorical_runs software
previously the categorical SAIGE results files were being combined with
`shared-source/combine_categorical_runs.out`. however, there was a precision error for Bonferroni
correction of small p-values. to fix this, i've migrated the code into my brand new shiny
cookiecutter-based C++ dev environment; and i've since implemented a fix using the high precision
library `mpfr`. for the time being, i'm just pulling in the dev copy of that software. i'll need to
iterate on that code a bit: there are other changes that really ought to be implemented. but once
that's frozen, i'll change this again to point to a configurable location