fix: fix ldscores workflow support for categorical output files and meta-analysis
ldscore regression support was only originally implemented for primary analysis output files. but
clearly people need that kind of information for other downstream output files, such as
meta-analysis for standard pipelines or the strange Bonferroni correction/min-p method for combining
SAIGE categorical analysis comparisons. this implements those extensions, which are arguably fixes
because as it was, things were only getting partially detected and incorrectly formatted in output.
this is a particularly unpleasant pipeline due to how divergent the categoricaldata analysis
folders' structure is. but what can you do, truly