feat: expose number of PCs in clean chip pipeline to userspace config
now the user should be able to set the number of smartpca-estimated PCs in Makefile.config. I need
to check to be sure that smartpca fastmode responds to this parameter, as there are some options in
the parameter file that are unceremoniously disabled in fastmode with no warning (e.g. outliers,
which i'm still including in the par file in case someone disables fastmode and wants to do outlier
removal like a good person)

re #10