fix: provide full paths to custom software for make tracking
some (not all) of the internal C++ programs used by the pipelines are tracked directly in rules.
this was good behavior for dev, probably worse now that things are slowing down. however,
regardless, that search process in make does not use the standard PATH search method. so the fastest
solution is to just provide the full installation path of the software, which when using conda is
just $(CONDA_PREFIX)/bin/*.out

re #1